# specs/001-cv-question-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CV Question Generator API
  description: |
    REST API for HR professionals to generate AI-powered interview questions 
    based on candidate CVs and job descriptions.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.cvquestiongen.com
    description: Production server

tags:
  - name: CVs
    description: CV management operations
  - name: Jobs
    description: Job description management operations
  - name: Questions
    description: Interview question generation

paths:
  /api/cvs:
    post:
      tags:
        - CVs
      summary: Upload and process CV
      description: |
        Upload unformatted CV text for AI extraction. The system will extract 
        structured data including skills with proficiency levels, work experience, 
        and education. Personal contact details are discarded after extraction.
        
        Uploading a new CV will overwrite any previously stored CV.
      operationId: uploadCv
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCvRequest'
            example:
              text: |
                John Doe
                Senior Software Engineer
                
                Skills: 5 years Java, 3 years Python, touched TypeScript
                
                Experience:
                - Senior Developer at TechCorp (2020-Present)
                  Built microservices using Spring Boot
                - Developer at StartupXYZ (2018-2020)
                  Full-stack development with React and Node.js
                
                Education:
                - BSc Computer Science, State University, 2018
      responses:
        '201':
          description: CV successfully uploaded and processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CvUploadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/AiUnavailable'

    get:
      tags:
        - CVs
      summary: Retrieve stored CV
      description: |
        Retrieve the currently stored CV including original text and 
        AI-extracted structured data.
      operationId: getCv
      responses:
        '200':
          description: CV retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CvResponse'
        '404':
          $ref: '#/components/responses/CvNotFound'

  /api/jobs:
    post:
      tags:
        - Jobs
      summary: Upload and process job description
      description: |
        Upload unformatted job description text for AI extraction. The system 
        will extract required skills, experience levels, and key competencies.
        
        Uploading a new job description will overwrite any previously stored one.
      operationId: uploadJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadJobRequest'
            example:
              text: |
                Senior Frontend Developer
                
                We are looking for an experienced frontend developer to join our team.
                
                Requirements:
                - 5+ years of experience with Angular
                - Strong TypeScript skills
                - Experience with state management (NgRx preferred)
                - Familiarity with CI/CD pipelines
                
                Nice to have:
                - React experience
                - Node.js backend knowledge
      responses:
        '201':
          description: Job description successfully uploaded and processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobUploadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/AiUnavailable'

    get:
      tags:
        - Jobs
      summary: Retrieve stored job description
      description: |
        Retrieve the currently stored job description including original text 
        and AI-extracted structured requirements.
      operationId: getJob
      responses:
        '200':
          description: Job description retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          $ref: '#/components/responses/JobNotFound'

  /api/questions:
    get:
      tags:
        - Questions
      summary: Generate interview questions
      description: |
        Generate interview questions based on the currently stored CV and job 
        description. Questions are grouped by topic and calibrated to the 
        candidate's proficiency level in each skill area.
        
        Each question includes answer guidelines for non-technical recruiters,
        key technical terms with explanations, and suggested follow-up questions.
        
        Requires both a CV and job description to be stored.
      operationId: generateQuestions
      responses:
        '200':
          description: Questions generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionsResponse'
        '400':
          description: CV or job description not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                cvMissing:
                  summary: CV not uploaded
                  value:
                    type: https://api.cvquestiongen.com/errors/cv-required
                    title: CV Required
                    status: 400
                    detail: No CV has been uploaded. Please upload a CV first using POST /api/cvs.
                    instance: /api/questions
                jobMissing:
                  summary: Job description not uploaded
                  value:
                    type: https://api.cvquestiongen.com/errors/job-required
                    title: Job Description Required
                    status: 400
                    detail: No job description has been uploaded. Please upload a job description first using POST /api/jobs.
                    instance: /api/questions
        '503':
          $ref: '#/components/responses/AiUnavailable'

components:
  schemas:
    # Request DTOs
    UploadCvRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 50000
          description: Unformatted CV text to process

    UploadJobRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 50000
          description: Unformatted job description text to process

    # Response DTOs
    CvUploadResponse:
      type: object
      required:
        - id
        - message
        - extraction
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the stored CV
        message:
          type: string
          description: Confirmation message
        extraction:
          $ref: '#/components/schemas/CvExtractionResult'

    CvResponse:
      type: object
      required:
        - id
        - originalText
        - extraction
        - extractedAt
      properties:
        id:
          type: string
          format: uuid
        originalText:
          type: string
          description: Original unformatted CV text
        extraction:
          $ref: '#/components/schemas/CvExtractionResult'
        extractedAt:
          type: string
          format: date-time
          description: When the CV was processed

    CvExtractionResult:
      type: object
      required:
        - personalInfo
        - skills
        - experience
        - education
      properties:
        personalInfo:
          $ref: '#/components/schemas/PersonalInfo'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        experience:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        warnings:
          type: array
          items:
            type: string
          description: Any warnings from extraction (e.g., no skills found)

    JobUploadResponse:
      type: object
      required:
        - id
        - message
        - extraction
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
        extraction:
          $ref: '#/components/schemas/JobExtractionResult'

    JobResponse:
      type: object
      required:
        - id
        - originalText
        - extraction
        - extractedAt
      properties:
        id:
          type: string
          format: uuid
        originalText:
          type: string
        extraction:
          $ref: '#/components/schemas/JobExtractionResult'
        extractedAt:
          type: string
          format: date-time

    JobExtractionResult:
      type: object
      required:
        - requiredSkills
        - competencies
        - experienceRequirements
      properties:
        requiredSkills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        competencies:
          type: array
          items:
            type: string
          description: Key competencies to assess
        experienceRequirements:
          type: array
          items:
            $ref: '#/components/schemas/ExperienceRequirement'
        warnings:
          type: array
          items:
            type: string

    QuestionsResponse:
      type: object
      required:
        - cvId
        - jobId
        - questionBlocks
        - generatedAt
      properties:
        cvId:
          type: string
          format: uuid
        jobId:
          type: string
          format: uuid
        questionBlocks:
          type: array
          items:
            $ref: '#/components/schemas/QuestionBlock'
        generatedAt:
          type: string
          format: date-time

    # Domain Models
    PersonalInfo:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        summary:
          type: string

    Skill:
      type: object
      required:
        - name
        - category
        - proficiency
      properties:
        name:
          type: string
          description: Skill name (e.g., "Java", "Angular")
        category:
          type: string
          enum:
            - Frontend
            - Backend
            - Database
            - DevOps
            - Cloud
            - Mobile
            - MachineLearning
            - Testing
            - SoftSkills
            - Other
        proficiency:
          type: string
          enum:
            - Beginner
            - Intermediate
            - Advanced
            - Expert
        yearsOfExperience:
          type: integer
          minimum: 0
        context:
          type: string
          description: Original text snippet for reference

    WorkExperience:
      type: object
      required:
        - jobTitle
        - company
      properties:
        jobTitle:
          type: string
        company:
          type: string
        duration:
          type: string
        responsibilities:
          type: array
          items:
            type: string

    Education:
      type: object
      required:
        - degree
        - institution
      properties:
        degree:
          type: string
        institution:
          type: string
        year:
          type: string

    ExperienceRequirement:
      type: object
      required:
        - area
        - minimumLevel
      properties:
        area:
          type: string
        minimumLevel:
          type: string
          enum:
            - Beginner
            - Intermediate
            - Advanced
            - Expert
        yearsRequired:
          type: integer

    QuestionBlock:
      type: object
      required:
        - topicLabel
        - difficulty
        - questions
      properties:
        topicLabel:
          type: string
          description: Topic label (e.g., "Frontend - Angular")
        difficulty:
          type: string
          enum:
            - Foundational
            - Standard
            - Advanced
        questions:
          type: array
          items:
            $ref: '#/components/schemas/InterviewQuestion'
          minItems: 5
          maxItems: 5

    InterviewQuestion:
      type: object
      required:
        - text
        - answerGuidelines
        - keyTerms
        - followUpQuestions
      properties:
        text:
          type: string
          description: The interview question
        answerGuidelines:
          type: string
          description: Non-technical explanation of what to look for in answers
        keyTerms:
          type: array
          items:
            $ref: '#/components/schemas/KeyTerm'
        followUpQuestions:
          type: array
          items:
            type: string

    KeyTerm:
      type: object
      required:
        - term
        - explanation
      properties:
        term:
          type: string
        explanation:
          type: string
          description: Plain English explanation

    # Error Models
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        retryAfter:
          type: integer
          description: Seconds to wait before retrying (for 503 errors)
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string

  responses:
    ValidationError:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc7231#section-6.5.1
            title: Validation Error
            status: 400
            detail: One or more validation errors occurred.
            errors:
              text:
                - The Text field is required.

    CvNotFound:
      description: CV not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://api.cvquestiongen.com/errors/cv-not-found
            title: CV Not Found
            status: 404
            detail: No CV has been uploaded yet.
            instance: /api/cvs

    JobNotFound:
      description: Job description not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://api.cvquestiongen.com/errors/job-not-found
            title: Job Description Not Found
            status: 404
            detail: No job description has been uploaded yet.
            instance: /api/jobs

    AiUnavailable:
      description: AI service unavailable
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://api.cvquestiongen.com/errors/ai-unavailable
            title: AI Service Unavailable
            status: 503
            detail: The AI service is temporarily unavailable. Please retry later.
            retryAfter: 30